Added support for custom parameters
